added groupController test mockBean InstructorService
Took 36 minutes

+++ src/main/java/io/codelex/studentsystem/EmployerController.java
@@ -3,7 +3,10 @@ package io.codelex.studentsystem;
 
 import io.codelex.studentsystem.api.Employer;
 import io.codelex.studentsystem.api.requests.AddEmployer;
+import io.codelex.studentsystem.repository.recordRepository.EmployerRecordRepository;
 import io.codelex.studentsystem.repository.service.EmployerService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -12,6 +15,7 @@ import javax.validation.Valid;
 @RequestMapping
 public class EmployerController {
     private final EmployerService employerService;
+    private EmployerRecordRepository employerRecordRepository;
 
     public EmployerController(EmployerService employerService) {
         this.employerService = employerService;
@@ -28,7 +32,12 @@ public class EmployerController {
     }
 
     @DeleteMapping("/internal-api/employers/{employerId}")
-    public void deleteById(@PathVariable long employerId) {
-        employerService.deleteById(employerId);
+    public ResponseEntity<String> deleteById(@PathVariable long employerId) {
+        if(employerService.findEmployerById(employerId) != null){
+            employerService.deleteById(employerId);
+            return new ResponseEntity<>(employerId + " Id was deleted", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Such id does not exist", HttpStatus.BAD_REQUEST);
+        }
     }
 }
